refactor: extract SceneObjectRouter for pure scene-object pass-throughs
Moves the six @Subscribe handlers for wall/decorative/ground object
spawn+despawn into a dedicated @Singleton SceneObjectRouter in the
lifecycle package. These handlers were pure pass-throughs to
ObjectHighlightOverlay with no other state or side-effects, so the
extraction is behavior-preserving.

The plugin registers/unregisters the router on the RuneLite EventBus
from startUp/shutDown to preserve @Subscribe lifecycle symmetry.

GameObjectSpawned/Despawned and ItemSpawned/Despawned are intentionally
left in the plugin because they interleave authoring-log side-effects
that touch plugin-private state (authoringLog, config, tick counters).
