+++ src/main/java/com/collectionloghelper/CollectionLogHelperPlugin.java
@@ -45,6 +45,7 @@ import com.collectionloghelper.efficiency.ScoredItem;
 import com.collectionloghelper.efficiency.SlayerStrategyCalculator;
 import com.collectionloghelper.guidance.GuidanceSequencer;
 import com.collectionloghelper.lifecycle.OverlayRegistry;
+import com.collectionloghelper.lifecycle.SceneObjectRouter;
 import com.collectionloghelper.overlay.CollectionLogWorldMapPoint;
 import com.collectionloghelper.overlay.DialogHighlightOverlay;
 import com.collectionloghelper.overlay.GroundItemHighlightOverlay;
@@ -81,15 +82,9 @@ import net.runelite.api.coords.WorldPoint;
 import net.runelite.api.events.ActorDeath;
 import net.runelite.api.events.AnimationChanged;
 import net.runelite.api.events.ChatMessage;
-import net.runelite.api.events.DecorativeObjectDespawned;
-import net.runelite.api.events.DecorativeObjectSpawned;
 import net.runelite.api.events.GameObjectDespawned;
 import net.runelite.api.events.GameObjectSpawned;
 import net.runelite.api.events.GameStateChanged;
-import net.runelite.api.events.GroundObjectDespawned;
-import net.runelite.api.events.GroundObjectSpawned;
-import net.runelite.api.events.WallObjectDespawned;
-import net.runelite.api.events.WallObjectSpawned;
 import net.runelite.api.events.GameTick;
 import net.runelite.api.events.HitsplatApplied;
 import net.runelite.api.events.InteractingChanged;
@@ -259,6 +254,9 @@ public class CollectionLogHelperPlugin extends Plugin
 	@Inject
 	private OverlayRegistry overlayRegistry;
 
+	@Inject
+	private SceneObjectRouter sceneObjectRouter;
+
 
 	private CollectionLogHelperPanel panel;
 	private NavigationButton navButton;
@@ -352,6 +350,7 @@ public class CollectionLogHelperPlugin extends Plugin
 			clientToolbar.addNavigation(navButton);
 		});
 		overlayRegistry.registerAll();
+		eventBus.register(sceneObjectRouter);
 
 		// If already logged in (e.g., plugin enabled mid-session), load state
 		if (client.getGameState() == GameState.LOGGED_IN)
@@ -392,6 +391,7 @@ public class CollectionLogHelperPlugin extends Plugin
 		}
 		clientToolbar.removeNavigation(navButton);
 		overlayRegistry.unregisterAll();
+		eventBus.unregister(sceneObjectRouter);
 		deactivateGuidance();
 		lastObtainedCount = -1;
 		sourcesWithMissingItems.clear();
@@ -2001,42 +2001,6 @@ public class CollectionLogHelperPlugin extends Plugin
 		WorldPoint wp = event.getTile().getWorldLocation();
 		authoringLog("OBJECT_DESPAWN id=%d at=[%d,%d,%d]",
 			event.getGameObject().getId(), wp.getX(), wp.getY(), wp.getPlane());
-	}
-
-	@Subscribe
-	public void onWallObjectSpawned(WallObjectSpawned event)
-	{
-		objectHighlightOverlay.onObjectSpawned(event.getWallObject());
-	}
-
-	@Subscribe
-	public void onWallObjectDespawned(WallObjectDespawned event)
-	{
-		objectHighlightOverlay.onObjectDespawned(event.getWallObject());
-	}
-
-	@Subscribe
-	public void onDecorativeObjectSpawned(DecorativeObjectSpawned event)
-	{
-		objectHighlightOverlay.onObjectSpawned(event.getDecorativeObject());
-	}
-
-	@Subscribe
-	public void onDecorativeObjectDespawned(DecorativeObjectDespawned event)
-	{
-		objectHighlightOverlay.onObjectDespawned(event.getDecorativeObject());
-	}
-
-	@Subscribe
-	public void onGroundObjectSpawned(GroundObjectSpawned event)
-	{
-		objectHighlightOverlay.onObjectSpawned(event.getGroundObject());
-	}
-
-	@Subscribe
-	public void onGroundObjectDespawned(GroundObjectDespawned event)
-	{
-		objectHighlightOverlay.onObjectDespawned(event.getGroundObject());
 	}
 
 	@Subscribe
